feat: modularize the entire codebase
