+++ src/utils/normalize.ts
@@ -1,0 +1,2 @@
+export const normalize = (u: string) =>
+  u.toLowerCase().trim().replace(/^@/, "");
